Elbow Assistive Embedded System
Integrated circuit that implements the device to assist the user to
position the elbow in a ranged position. The embedded system involves a
linear actuator which will position the elbow through the use of EMG
sensors and an artificial neural network using pattern recognition which
is used to map the EMG readings. Programmed an Intel Edison using C++ to
control the linear actuator by positioning them to the desired position.